Fix: Adiciona algumas funções no server.js e atualização no package.json dos módulos novos adicionados
