--- config/db.js
@@ -1,18 +1,0 @@
-const { Pool } = require('pg');
-require('dotenv').config();
-
-const isSSL = process.env.DB_SSL === 'true';
-
-const pool = new Pool({
-    user: process.env.DB_USER,
-    host: process.env.DB_HOST,
-    database: process.env.DB_DATABASE,
-    password: process.env.DB_PASSWORD,
-    port: process.env.DB_PORT,
-    ssl: isSSL ? { rejectUnauthorized: false } : false,
-  });
-
-module.exports = {
-  query: (text, params) => pool.query(text, params),
-  connect: () => pool.connect(),
-};
